refactor: rename close-channel protocol to channel-closer protocol
